Changing float to halfs in the SSAO shaders. Also fixing an issue with fallback shader name in ParticlesLit.shader.

+++ com.unity.render-pipelines.universal/ShaderLibrary/SSAO.hlsl
@@ -15,14 +15,14 @@ SAMPLER(sampler_BaseMap);
 SAMPLER(sampler_ScreenSpaceOcclusionTexture);
 
 // Params
-float4 _SSAOParams;
-float4 _SourceSize;
-float4 _ProjectionParams2;
-float4x4 _CameraViewProjections[2]; // This is different from UNITY_MATRIX_VP (platform-agnostic projection matrix is used). Handle both non-XR and XR modes.
-float4 _CameraViewTopLeftCorner[2]; // TODO: check if we can use half type
-float4 _CameraViewXExtent[2];
-float4 _CameraViewYExtent[2];
-float4 _CameraViewZExtent[2];
+half4 _SSAOParams;
+half4 _SourceSize;
+half4 _ProjectionParams2;
+half4x4 _CameraViewProjections[2]; // This is different from UNITY_MATRIX_VP (platform-agnostic projection matrix is used). Handle both non-XR and XR modes.
+half4 _CameraViewTopLeftCorner[2];
+half4 _CameraViewXExtent[2];
+half4 _CameraViewYExtent[2];
+half4 _CameraViewZExtent[2];
 
 // SSAO Settings
 #define INTENSITY _SSAOParams.x
@@ -45,16 +45,16 @@ float4 _CameraViewZExtent[2];
 // Constants
 // kContrast determines the contrast of occlusion. This allows users to control over/under
 // occlusion. At the moment, this is not exposed to the editor because it's rarely useful.
-static const float kContrast = 0.6;
+static const half kContrast = 0.6;
 
 // The constant below controls the geometry-awareness of the bilateral
 // filter. The higher value, the more sensitive it is.
-static const float kGeometryCoeff = 0.8;
+static const half kGeometryCoeff = 0.8;
 
 // The constants below are used in the AO estimator. Beta is mainly used for suppressing
 // self-shadowing noise, and Epsilon is used to prevent calculation underflow. See the
 // paper (Morgan 2011 http://goo.gl/2iz3P) for further details of these constants.
-static const float kBeta = 0.002;
+static const half kBeta = 0.002;
 #define EPSILON         1.0e-4
 
 #if defined(USING_STEREO_MATRICES)
@@ -63,22 +63,22 @@ static const float kBeta = 0.002;
 #define unity_eyeIndex 0
 #endif
 
-float4 PackAONormal(float ao, float3 n)
+half4 PackAONormal(half ao, half3 n)
 {
-    return float4(ao, n * 0.5 + 0.5);
+    return half4(ao, n * 0.5 + 0.5);
 }
 
-float3 GetPackedNormal(float4 p)
+half3 GetPackedNormal(half4 p)
 {
     return p.gba * 2.0 - 1.0;
 }
 
-float GetPackedAO(float4 p)
+half GetPackedAO(half4 p)
 {
     return p.r;
 }
 
-float EncodeAO(float x)
+half EncodeAO(half x)
 {
     #if UNITY_COLORSPACE_GAMMA
         return 1.0 - max(LinearToSRGB(1.0 - saturate(x)), 0.0);
@@ -87,42 +87,42 @@ float EncodeAO(float x)
     #endif
 }
 
-float CompareNormal(float3 d1, float3 d2)
+half CompareNormal(half3 d1, half3 d2)
 {
     return smoothstep(kGeometryCoeff, 1.0, dot(d1, d2));
 }
 
-float2 GetScreenSpacePosition(float2 uv)
+half2 GetScreenSpacePosition(half2 uv)
 {
     return uv * SCREEN_PARAMS.xy * DOWNSAMPLE;
 }
 
 // Trigonometric function utility
-float2 CosSin(float theta)
+half2 CosSin(half theta)
 {
-    float sn, cs;
+    half sn, cs;
     sincos(theta, sn, cs);
-    return float2(cs, sn);
+    return half2(cs, sn);
 }
 
 // Pseudo random number generator with 2D coordinates
-float UVRandom(float u, float v)
+half UVRandom(half u, half v)
 {
-    float f = dot(float2(12.9898, 78.233), float2(u, v));
+    half f = dot(half2(12.9898, 78.233), half2(u, v));
     return frac(43758.5453 * sin(f));
 }
 
 // Sample point picker
-float3 PickSamplePoint(float2 uv, float randAddon, int index)
+half3 PickSamplePoint(half2 uv, half randAddon, int index)
 {
-    float2 positionSS = GetScreenSpacePosition(uv);
-    float gn = InterleavedGradientNoise(positionSS, index);
-    float u = frac(UVRandom(0.0, index + randAddon) + gn) * 2.0 - 1.0;
-    float theta = (UVRandom(1.0, index + randAddon) + gn) * TWO_PI;
-    return float3(CosSin(theta) * sqrt(1.0 - u * u), u);
+    half2 positionSS = GetScreenSpacePosition(uv);
+    half gn = InterleavedGradientNoise(positionSS, index);
+    half u = frac(UVRandom(0.0, index + randAddon) + gn) * 2.0 - 1.0;
+    half theta = (UVRandom(1.0, index + randAddon) + gn) * TWO_PI;
+    return half3(CosSin(theta) * sqrt(1.0 - u * u), u);
 }
 
-float RawToLinearDepth(float rawDepth)
+half RawToLinearDepth(half rawDepth)
 {
     #if defined(_ORTHOGRAPHIC)
         #if UNITY_REVERSED_Z
@@ -135,28 +135,28 @@ float RawToLinearDepth(float rawDepth)
     #endif
 }
 
-float SampleAndGetLinearDepth(float2 uv)
+half SampleAndGetLinearDepth(half2 uv)
 {
-    float rawDepth = SampleSceneDepth(uv.xy).r;
+    half rawDepth = SampleSceneDepth(uv.xy).r;
     return RawToLinearDepth(rawDepth);
 }
 
 // This returns a vector in world unit (not a position), from camera to the given point described by uv screen coordinate and depth (in absolute world unit).
-float3 ReconstructViewPos(float2 uv, float depth)
+half3 ReconstructViewPos(half2 uv, half depth)
 {
     // Screen is y-inverted.
     uv.y = 1.0 - uv.y;
 
     // view pos in world space
     #if defined(_ORTHOGRAPHIC)
-        float zScale = depth * _ProjectionParams.w; // divide by far plane
-        float3 viewPos = _CameraViewTopLeftCorner[unity_eyeIndex].xyz
+        half zScale = depth * _ProjectionParams.w; // divide by far plane
+        half3 viewPos = _CameraViewTopLeftCorner[unity_eyeIndex].xyz
                        + _CameraViewXExtent[unity_eyeIndex].xyz * uv.x
                        + _CameraViewYExtent[unity_eyeIndex].xyz * uv.y
                        + _CameraViewZExtent[unity_eyeIndex].xyz * zScale;
     #else
-        float zScale = depth * _ProjectionParams2.x; // divide by near plane
-        float3 viewPos = _CameraViewTopLeftCorner[unity_eyeIndex].xyz
+        half zScale = depth * _ProjectionParams2.x; // divide by near plane
+        half3 viewPos = _CameraViewTopLeftCorner[unity_eyeIndex].xyz
                        + _CameraViewXExtent[unity_eyeIndex].xyz * uv.x
                        + _CameraViewYExtent[unity_eyeIndex].xyz * uv.y;
         viewPos *= zScale;
@@ -171,23 +171,23 @@ float3 ReconstructViewPos(float2 uv, float depth)
 // High:   5 taps on each direction: | z | x | * | y | w |
 // https://atyuwen.github.io/posts/normal-reconstruction/
 // https://wickedengine.net/2019/09/22/improved-normal-reconstruction-from-depth/
-float3 ReconstructNormal(float2 uv, float depth, float3 vpos)
+half3 ReconstructNormal(half2 uv, half depth, half3 vpos)
 {
     #if defined(_RECONSTRUCT_NORMAL_LOW)
         return normalize(cross(ddy(vpos), ddx(vpos)));
     #else
-        float2 delta = _SourceSize.zw * 2.0;
+        half2 delta = _SourceSize.zw * 2.0;
 
         // Sample the neighbour fragments
-        float2 lUV = float2(-delta.x, 0.0);
-        float2 rUV = float2( delta.x, 0.0);
-        float2 uUV = float2(0.0,  delta.y);
-        float2 dUV = float2(0.0, -delta.y);
+        half2 lUV = half2(-delta.x, 0.0);
+        half2 rUV = half2( delta.x, 0.0);
+        half2 uUV = half2(0.0,  delta.y);
+        half2 dUV = half2(0.0, -delta.y);
 
-        float3 l1 = float3(uv + lUV, 0.0); l1.z = SampleAndGetLinearDepth(l1.xy); // Left1
-        float3 r1 = float3(uv + rUV, 0.0); r1.z = SampleAndGetLinearDepth(r1.xy); // Right1
-        float3 u1 = float3(uv + uUV, 0.0); u1.z = SampleAndGetLinearDepth(u1.xy); // Up1
-        float3 d1 = float3(uv + dUV, 0.0); d1.z = SampleAndGetLinearDepth(d1.xy); // Down1
+        half3 l1 = half3(uv + lUV, 0.0); l1.z = SampleAndGetLinearDepth(l1.xy); // Left1
+        half3 r1 = half3(uv + rUV, 0.0); r1.z = SampleAndGetLinearDepth(r1.xy); // Right1
+        half3 u1 = half3(uv + uUV, 0.0); u1.z = SampleAndGetLinearDepth(u1.xy); // Up1
+        half3 d1 = half3(uv + dUV, 0.0); d1.z = SampleAndGetLinearDepth(d1.xy); // Down1
 
         // Determine the closest horizontal and vertical pixels...
         // horizontal: left = 0.0 right = 1.0
@@ -196,10 +196,10 @@ float3 ReconstructNormal(float2 uv, float depth, float3 vpos)
              uint closest_horizontal = l1.z > r1.z ? 0 : 1;
              uint closest_vertical   = d1.z > u1.z ? 0 : 1;
         #else
-            float3 l2 = float3(uv + lUV * 2.0, 0.0); l2.z = SampleAndGetLinearDepth(l2.xy); // Left2
-            float3 r2 = float3(uv + rUV * 2.0, 0.0); r2.z = SampleAndGetLinearDepth(r2.xy); // Right2
-            float3 u2 = float3(uv + uUV * 2.0, 0.0); u2.z = SampleAndGetLinearDepth(u2.xy); // Up2
-            float3 d2 = float3(uv + dUV * 2.0, 0.0); d2.z = SampleAndGetLinearDepth(d2.xy); // Down2
+            half3 l2 = half3(uv + lUV * 2.0, 0.0); l2.z = SampleAndGetLinearDepth(l2.xy); // Left2
+            half3 r2 = half3(uv + rUV * 2.0, 0.0); r2.z = SampleAndGetLinearDepth(r2.xy); // Right2
+            half3 u2 = half3(uv + uUV * 2.0, 0.0); u2.z = SampleAndGetLinearDepth(u2.xy); // Up2
+            half3 d2 = half3(uv + dUV * 2.0, 0.0); d2.z = SampleAndGetLinearDepth(d2.xy); // Down2
 
             const uint closest_horizontal = abs( (2.0 * l1.z - l2.z) - depth) < abs( (2.0 * r1.z - r2.z) - depth) ? 0 : 1;
             const uint closest_vertical   = abs( (2.0 * d1.z - d2.z) - depth) < abs( (2.0 * u1.z - u2.z) - depth) ? 0 : 1;
@@ -213,8 +213,8 @@ float3 ReconstructNormal(float2 uv, float depth, float3 vpos)
         // h == 1.0 && v == 1.0: p1 = right, p2 = up
         // h == 0.0 && v == 1.0: p1 = up,    p2 = left
         // Calculate the view space positions for the three points...
-        float3 P1;
-        float3 P2;
+        half3 P1;
+        half3 P2;
         if (closest_vertical == 0)
         {
             P1 = closest_horizontal == 0 ? l1 : d1;
@@ -234,7 +234,7 @@ float3 ReconstructNormal(float2 uv, float depth, float3 vpos)
     #endif
 }
 
-void SampleDepthNormalView(float2 uv, out float depth, out float3 normal, out float3 vpos)
+void SampleDepthNormalView(half2 uv, out half depth, out half3 normal, out half3 vpos)
 {
     depth  = SampleAndGetLinearDepth(uv);
     vpos = ReconstructViewPos(uv, depth);
@@ -249,26 +249,26 @@ void SampleDepthNormalView(float2 uv, out float depth, out float3 normal, out fl
 // Distance-based AO estimator based on Morgan 2011
 // "Alchemy screen-space ambient obscurance algorithm"
 // http://graphics.cs.williams.edu/papers/AlchemyHPG11/
-float4 SSAO(Varyings input) : SV_Target
+half4 SSAO(Varyings input) : SV_Target
 {
     UNITY_SETUP_STEREO_EYE_INDEX_POST_VERTEX(input);
-    float2 uv = input.uv;
+    half2 uv = input.uv;
 
     // Parameters used in coordinate conversion
-    float2 p11_22, p13_31;
-    float3x3 camTransform = (float3x3)_CameraViewProjections[unity_eyeIndex]; // camera viewProjection matrix
+    half2 p11_22, p13_31;
+    half3x3 camTransform = (half3x3)_CameraViewProjections[unity_eyeIndex]; // camera viewProjection matrix
 
     // Get the depth, normal and view position for this fragment
-    float depth_o;
-    float3 norm_o;
-    float3 vpos_o;
+    half depth_o;
+    half3 norm_o;
+    half3 vpos_o;
     SampleDepthNormalView(uv, depth_o, norm_o, vpos_o);
 
     // This was added to avoid a NVIDIA driver issue.
-    float randAddon = uv.x * 1e-10;
+    half randAddon = uv.x * 1e-10;
 
-    float rcpSampleCount = rcp(SAMPLE_COUNT);
-    float ao = 0.0;
+    half rcpSampleCount = rcp(SAMPLE_COUNT);
+    half ao = 0.0;
     for (int s = 0; s < int(SAMPLE_COUNT); s++)
     {
         #if defined(SHADER_API_D3D11)
@@ -277,34 +277,34 @@ float4 SSAO(Varyings input) : SV_Target
         #endif
 
         // Sample point
-        float3 v_s1 = PickSamplePoint(uv, randAddon, s); // (kchang) should we rotate this "random" vector to world space?
+        half3 v_s1 = PickSamplePoint(uv, randAddon, s); // (kchang) should we rotate this "random" vector to world space?
 
         // Make it distributed between [0, _Radius]
         v_s1 *= sqrt((s + 1.0) * rcpSampleCount ) * RADIUS;
 
         v_s1 = faceforward(v_s1, -norm_o, v_s1);
-        float3 vpos_s1 = vpos_o + v_s1;
+        half3 vpos_s1 = vpos_o + v_s1;
 
         // Reproject the sample point
-        float3 spos_s1 = mul(camTransform, vpos_s1);
+        half3 spos_s1 = mul(camTransform, vpos_s1);
 
         #if defined(_ORTHOGRAPHIC)
-            float2 uv_s1_01 = clamp((spos_s1.xy + 1.0) * 0.5, 0.0, 1.0);
+            half2 uv_s1_01 = clamp((spos_s1.xy + 1.0) * 0.5, 0.0, 1.0);
         #else
-            float zdist = -dot(UNITY_MATRIX_V[2].xyz, vpos_s1);
-            float2 uv_s1_01 = clamp((spos_s1.xy * rcp(zdist) + 1.0) * 0.5, 0.0, 1.0);
+            half zdist = -dot(UNITY_MATRIX_V[2].xyz, vpos_s1);
+            half2 uv_s1_01 = clamp((spos_s1.xy * rcp(zdist) + 1.0) * 0.5, 0.0, 1.0);
         #endif
 
         // Depth at the sample point
-        float depth_s1 = SampleAndGetLinearDepth(uv_s1_01);
+        half depth_s1 = SampleAndGetLinearDepth(uv_s1_01);
 
         // Relative position of the sample point
-        float3 vpos_s2 = ReconstructViewPos(uv_s1_01, depth_s1);
-        float3 v_s2 = vpos_s2 - vpos_o;
+        half3 vpos_s2 = ReconstructViewPos(uv_s1_01, depth_s1);
+        half3 v_s2 = vpos_s2 - vpos_o;
 
         // Estimate the obscurance value
-        float a1 = max(dot(v_s2, norm_o) - kBeta * depth_o, 0.0);
-        float a2 = dot(v_s2, v_s2) + EPSILON;
+        half a1 = max(dot(v_s2, norm_o) - kBeta * depth_o, 0.0);
+        half a2 = dot(v_s2, v_s2) + EPSILON;
         ao += a1 * rcp(a2);
     }
 
@@ -317,35 +317,35 @@ float4 SSAO(Varyings input) : SV_Target
 }
 
 // Geometry-aware separable bilateral filter
-half4 Blur(float2 uv, float2 delta) : SV_Target
+half4 Blur(half2 uv, half2 delta) : SV_Target
 {
-    float4 p0 = SAMPLE_BASEMAP(uv                 );
-    float4 p1a = SAMPLE_BASEMAP(uv - delta * 1.3846153846);
-    float4 p1b = SAMPLE_BASEMAP(uv + delta * 1.3846153846);
-    float4 p2a = SAMPLE_BASEMAP(uv - delta * 3.2307692308);
-    float4 p2b = SAMPLE_BASEMAP(uv + delta * 3.2307692308);
+    half4 p0 = SAMPLE_BASEMAP(uv                 );
+    half4 p1a = SAMPLE_BASEMAP(uv - delta * 1.3846153846);
+    half4 p1b = SAMPLE_BASEMAP(uv + delta * 1.3846153846);
+    half4 p2a = SAMPLE_BASEMAP(uv - delta * 3.2307692308);
+    half4 p2b = SAMPLE_BASEMAP(uv + delta * 3.2307692308);
 
     #if defined(BLUR_SAMPLE_CENTER_NORMAL)
         #if defined(_SOURCE_DEPTH_NORMALS)
-            float3 n0 = SampleSceneNormals(uv);
+            half3 n0 = SampleSceneNormals(uv);
         #else
             // Get the depth, normal and view position for this fragment
-            float depth_o;
-            float3 n0;
-            float3 vpos_o;
+            half depth_o;
+            half3 n0;
+            half3 vpos_o;
             SampleDepthNormalView(uv, depth_o, n0, vpos_o);
         #endif
     #else
-        float3 n0 = GetPackedNormal(p0);
+        half3 n0 = GetPackedNormal(p0);
     #endif
 
-    float w0  =                                           0.2270270270;
-    float w1a = CompareNormal(n0, GetPackedNormal(p1a)) * 0.3162162162;
-    float w1b = CompareNormal(n0, GetPackedNormal(p1b)) * 0.3162162162;
-    float w2a = CompareNormal(n0, GetPackedNormal(p2a)) * 0.0702702703;
-    float w2b = CompareNormal(n0, GetPackedNormal(p2b)) * 0.0702702703;
+    half w0  =                                           0.2270270270;
+    half w1a = CompareNormal(n0, GetPackedNormal(p1a)) * 0.3162162162;
+    half w1b = CompareNormal(n0, GetPackedNormal(p1b)) * 0.3162162162;
+    half w2a = CompareNormal(n0, GetPackedNormal(p2a)) * 0.0702702703;
+    half w2b = CompareNormal(n0, GetPackedNormal(p2b)) * 0.0702702703;
 
-    float s;
+    half s;
     s  = GetPackedAO(p0)  * w0;
     s += GetPackedAO(p1a) * w1a;
     s += GetPackedAO(p1b) * w1b;
@@ -358,23 +358,23 @@ half4 Blur(float2 uv, float2 delta) : SV_Target
 }
 
 // Geometry-aware bilateral filter (single pass/small kernel)
-float BlurSmall(float2 uv, float2 delta)
+half BlurSmall(half2 uv, half2 delta)
 {
-    float4 p0 = SAMPLE_BASEMAP(uv                             );
-    float4 p1 = SAMPLE_BASEMAP(uv + float2(-delta.x, -delta.y));
-    float4 p2 = SAMPLE_BASEMAP(uv + float2( delta.x, -delta.y));
-    float4 p3 = SAMPLE_BASEMAP(uv + float2(-delta.x,  delta.y));
-    float4 p4 = SAMPLE_BASEMAP(uv + float2( delta.x,  delta.y));
+    half4 p0 = SAMPLE_BASEMAP(uv                             );
+    half4 p1 = SAMPLE_BASEMAP(uv + half2(-delta.x, -delta.y));
+    half4 p2 = SAMPLE_BASEMAP(uv + half2( delta.x, -delta.y));
+    half4 p3 = SAMPLE_BASEMAP(uv + half2(-delta.x,  delta.y));
+    half4 p4 = SAMPLE_BASEMAP(uv + half2( delta.x,  delta.y));
 
-    float3 n0 = GetPackedNormal(p0);
+    half3 n0 = GetPackedNormal(p0);
 
-    float w0 = 1.0;
-    float w1 = CompareNormal(n0, GetPackedNormal(p1));
-    float w2 = CompareNormal(n0, GetPackedNormal(p2));
-    float w3 = CompareNormal(n0, GetPackedNormal(p3));
-    float w4 = CompareNormal(n0, GetPackedNormal(p4));
+    half w0 = 1.0;
+    half w1 = CompareNormal(n0, GetPackedNormal(p1));
+    half w2 = CompareNormal(n0, GetPackedNormal(p2));
+    half w3 = CompareNormal(n0, GetPackedNormal(p3));
+    half w4 = CompareNormal(n0, GetPackedNormal(p4));
 
-    float s;
+    half s;
     s  = GetPackedAO(p0) * w0;
     s += GetPackedAO(p1) * w1;
     s += GetPackedAO(p2) * w2;
@@ -388,8 +388,8 @@ half4 HorizontalBlur(Varyings input) : SV_Target
 {
     UNITY_SETUP_STEREO_EYE_INDEX_POST_VERTEX(input);
 
-    float2 uv = input.uv;
-    float2 delta = float2(_SourceSize.z, 0.0);
+    half2 uv = input.uv;
+    half2 delta = half2(_SourceSize.z, 0.0);
     return Blur(uv, delta);
 }
 
@@ -397,8 +397,8 @@ half4 VerticalBlur(Varyings input) : SV_Target
 {
     UNITY_SETUP_STEREO_EYE_INDEX_POST_VERTEX(input);
 
-    float2 uv = input.uv;
-    float2 delta = float2(0.0, _SourceSize.w);
+    half2 uv = input.uv;
+    half2 delta = half2(0.0, _SourceSize.w);
     return Blur(uv, delta);
 }
 
@@ -406,8 +406,8 @@ half4 FinalBlur(Varyings input) : SV_Target
 {
     UNITY_SETUP_STEREO_EYE_INDEX_POST_VERTEX(input);
 
-    float2 uv = input.uv;
-    float2 delta = _SourceSize.zw;
+    half2 uv = input.uv;
+    half2 delta = _SourceSize.zw;
     return 1.0 - BlurSmall(uv, delta);
 }
